requirements.txt
djangorestframework
djangorestframework-simplejwt
Django==4.0.3